build: refactor build a bit to work on OS/X
Moved profile+status to runtime for now.
Moved library dependencies from wake to runtime.
